split solar into a library and cli by making the repo into a workspace
