+++ solar/src/actors/network/lan_discovery.rs
