+++ solar/src/error.rs
