+++ solar/src/storage/blob.rs
