+++ solar_cli/build.rs
@@ -1,5 +1,6 @@
 // build.rs
 use std::process::Command;
+
 fn main() {
     let git_hash = if let Ok(output) = Command::new("git").args(["rev-parse", "HEAD"]).output() {
         String::from_utf8(output.stdout).unwrap()
